Write test to MovieListModule
